requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
docopt==0.6.2
gunicorn==20.1.0
num2words==0.5.12
Pillow==9.2.0
psycopg2-binary==2.9.4
pydantic==1.10.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.5
whitenoise==6.2.0
